+++ src/middlewares/transmit/compression.js
@@ -7,6 +7,7 @@
 "use strict";
 
 const { defaultsDeep } = require("lodash");
+const { parseByteString } = require("../../utils");
 const zlib = require("zlib");
 const Promise = require("bluebird");
 
@@ -17,9 +18,10 @@ const Promise = require("bluebird");
  * @param {String} method
  */
 module.exports = function CompressionMiddleware(opts) {
-	opts = defaultsDeep(opts, { method: "deflate" });
+	opts = defaultsDeep(opts, { method: "deflate", threshold: "1kb" });
 
 	let compress, decompress;
+	const threshold = parseByteString(opts.threshold);
 
 	switch(opts.method) {
 		case "deflate":
@@ -39,31 +41,44 @@ module.exports = function CompressionMiddleware(opts) {
 			throw new Error("Unknow compression method: " + opts.method);
 	}
 
+	let logger;
+
 	return {
 		name: "Compression",
 
-		created() {
+		created(broker) {
+			logger = broker.getLogger("TX-COMPRESS");
 			/* istanbul ignore next */
-			this.logger.info(`The transmission is COMPRESSED by '${opts.method}'.`);
+			logger.info(`The transmission is COMPRESSED by '${opts.method}'. Threshold: ${ threshold != null ? opts.threshold : "none"}`);
 		},
 
 		transporterSend(next) {
 			return (topic, data, meta) => {
+				if (threshold != null && data.length < threshold) {
+					logger.debug(`Packet '${topic}' is small and not compressed. Size: ${data.length}`);
+					return next(topic, Buffer.concat([Buffer.from([0x00]), data]), meta);
+				}
 				return compress(data)
 					.then(res => {
-						this.logger.debug(`Packet '${topic}' compressed. Saving: ${Number((1 - (res.length / data.length)) * 100).toFixed(0)}%`, data.length, res.length);
-						return next(topic, res, meta);
+						logger.debug(`Packet '${topic}' compressed. Saving: ${Number((1 - (res.length / data.length)) * 100).toFixed(0)}%`, data.length, res.length);
+						return next(topic, Buffer.concat([Buffer.from([0x01]), res]), meta);
 					});
 			};
 		},
 
 		transporterReceive(next) {
 			return (cmd, data, s) => {
-				return decompress(data)
-					.then(res => {
-						this.logger.debug(`Packet '${cmd}' decompressed. Saving: ${Number((1 - (res.length / data.length)) * 100).toFixed(0)}%`, data.length, res.length);
-						return next(cmd, res, s);
-					});
+				const isCompressed = data.readInt8(0);
+				if (isCompressed == 0) {
+					logger.debug(`Packet '${cmd}' is small and not compressed. Size: ${data.length}`);
+					return next(cmd, data.slice(1), s);
+				} else {
+					return decompress(data.slice(1))
+						.then(res => {
+							logger.debug(`Packet '${cmd}' decompressed. Saving: ${Number((1 - (data.length / res.length)) * 100).toFixed(0)}%`, res.length, data.length);
+							return next(cmd, res, s);
+						});
+				}
 			};
 		}
 	};